Changed type snpz to long

+++ src/main/java/org/acme/foodpackaging/domain/PackagingSchedule.java
@@ -36,11 +36,11 @@ public class PackagingSchedule {
     @ValueRangeProvider
     private List<Job> jobs;
 
-    private Map<Integer, DbJobRow> dbJobRowMap;
+    private Map<Long, DbJobRow> dbJobRowMap;
 
     private Map<Long, DbMaintenanceRow> dbMaintenanceRowMap;
 
-    private  Map<Integer, Job> jobIdMap;
+    private  Map<Long, Job> jobIdMap;
 
     @PlanningScore
     private HardMediumSoftLongScore score;
